Hopefully fix building on Windows
