+++ src/core/model_private_macros.h
@@ -1,4 +1,4 @@
-/* model/ship.h: Class representation of a ship.
+/* core/model_private_macros.h: Utility macros used for constructing objects from ASTs.
  *
  * Copyright 2019 Adrian "ArdiMaster" Welcker
  *
@@ -15,36 +15,13 @@
  * limitations under the License.
  */
 
-
 #pragma once
 
-#ifndef STELLARIS_STAT_VIEWER_SHIP_H
-#define STELLARIS_STAT_VIEWER_SHIP_H
+#ifndef STELLARIS_STAT_VIEWER_MODEL_PRIVATE_MACROS_H
+#define STELLARIS_STAT_VIEWER_MODEL_PRIVATE_MACROS_H
 
-#include <QtCore/QObject>
+#define CHECK_PTR(ptr) do { if (!(ptr)) { delete state; return nullptr; } } while (0)
+#define CHECK_COMPOUND(node) do { if (!(node) || (node)->type != Parsing::NT_COMPOUND) { delete state; return nullptr; } } while (0)
+#define ITERATE_CHILDREN(node, cn) for (AstNode *cn = (node)->val.firstChild; cn; cn = cn->nextSibling)
 
-namespace Parsing { struct AstNode; }
-
-namespace Galaxy {
-	class Fleet;
-	class ShipDesign;
-	class State;
-
-	class Ship : public QObject {
-		Q_OBJECT
-	public:
-		Ship(State *parent);
-		qint64 getIndex() const;
-		const QString &getName() const;
-		Fleet *getFleet() const;
-		ShipDesign *getDesign() const;
-		static Ship *createFromAst(Parsing::AstNode *tree, State *parent);
-	private:
-		qint64 index;
-		QString name;
-		Fleet *fleet;
-		ShipDesign *design;
-	};
-}
-
-#endif
+#endif //STELLARIS_STAT_VIEWER_MODEL_PRIVATE_MACROS_H
