+++ src/core/ssv_core.h
@@ -1,0 +1,18 @@
+//
+// Created by Adrian Welcker on 2019-06-04.
+//
+
+#ifndef STELLARIS_STAT_VIEWER_SSV_CORE_H
+#define STELLARIS_STAT_VIEWER_SSV_CORE_H
+
+#ifdef _MSC_VER
+#ifdef LIBRARY_BUILD
+#define DLLEXPORT __declspec(dllexport)
+#else
+#define DLLEXPORT __declspec(dllimport)
+#endif
+#else
+#define DLLEXPORT
+#endif
+
+#endif //STELLARIS_STAT_VIEWER_SSV_CORE_H
